+++ build.gradle
@@ -1,16 +1,19 @@
 plugins {
     id 'java'
     id 'application'
+    id 'checkstyle'
     id 'com.github.johnrengelman.shadow' version '7.1.2'
 }
 
 repositories {
     mavenCentral()
+    maven { url 'https://oss.sonatype.org/content/repositories/snapshots/' }
 }
 
 dependencies {
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.10.0'
     testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.10.0'
+    implementation group: 'com.joestelmach', name: 'natty', version: '0.6'
 }
 
 test {
@@ -28,7 +31,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("duke.Duke")
+}
+
+checkstyle {
+    toolVersion = '10.2'
 }
 
 shadowJar {
@@ -39,4 +46,6 @@ shadowJar {
 
 run{
     standardInput = System.in
+    enableAssertions = true
 }
+
