+++ src/main/java/duke/exception/MissingTimeException.java
@@ -2,6 +2,6 @@ package duke.exception;
 
 public class MissingTimeException extends DukeException {
     public MissingTimeException() {
-        super("Please input a date or time for this task.");
+        super("Please input a date and time for this task.");
     }
 }
