+++ src/main/java/duke/utility/TaskList.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.utility;
 
 import duke.exception.InvalidTaskException;
 import duke.task.Task;
