+++ text-ui-test/data/task.txt
@@ -1,3 +1,4 @@
-T/@/1/@/finish oop
-D/@/1/@/indiv project /@/2023-09-01T18:00
-E/@/1/@/bobi birthday /@/2023-05-30T18:00/@/2023-05-30T23:00
+T/@/1/@/say hello
+D/@/0/@/find bobi /@/2023-09-02T23:59
+E/@/0/@/bobi birthday /@/2023-09-03T18:00/@/2023-03-02T18:00
+D/@/0/@/finish junit and javadoc /@/2023-09-02T18:00
